Fix warning on Leopard, tested ok on Tiger: vo_macosx.m: In function '-[MPlayerOpenGLView config]': vo_macosx.m:387: warning: passing argument 1 of 'setValues:forParameter:' from incompatible pointer type
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@25165 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libvo/vo_macosx.m
@@ -370,7 +370,7 @@ static int control(uint32_t request, void *data, ...)
 	uint32_t d_width;
 	uint32_t d_height;
 	
-	long swapInterval = 1;
+	GLint swapInterval = 1;
 	
 	NSRect frame;
 	CVReturn error = kCVReturnSuccess;
